remove options service dependency
