+++ research-hub-web/src/app/components/content-details/content-details.component.ts
@@ -5,11 +5,32 @@ import { Content } from '@model/Content';
 import { Location } from '@angular/common';
 import { AnalyticsService } from '@services/analytics.service';
 import { ListItem } from '@model/ListItem';
-import { ActionTypeId, ContentTypeId, RoleTypeId } from '@services/options.service';
 import { Subscription } from 'rxjs';
 import { MediaChange, MediaObserver } from '@angular/flex-layout';
 import { LayoutService } from '@services/layout.service';
 import { AppComponentService } from '@app/app.component.service';
+
+enum ActionTypeId {
+  Integrated = 1,
+  ExternalUrl,
+  ExternalMailTo
+}
+
+enum RoleTypeId {
+  ServiceOwner = 1,
+  CopyOwner,
+  UserSupport
+}
+
+enum ContentTypeId {
+  Support = 1,
+  Equipment,
+  Training,
+  Software,
+  Facilities,
+  KnowledgeArticle,
+  Guide
+}
 
 @Component({
   selector: 'app-content-details',
