+++ research-hub-web/src/app/components/home/browse/browse.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit, Input, OnDestroy } from '@angular/core';
-import { OptionsService } from '@services/options.service';
 import { Subscription } from 'rxjs';
 import { MediaChange, MediaObserver } from '@angular/flex-layout';
 import { LayoutService } from '@services/layout.service';
@@ -10,6 +9,8 @@ import { LayoutService } from '@services/layout.service';
   styleUrls: ['./browse.component.scss']
 })
 export class BrowseComponent implements OnInit, OnDestroy {
+  public categoryOptions: any[];
+  public categoryOptionsGQL: any[];
 
   private mediaSub: Subscription;
 
@@ -22,7 +23,7 @@ export class BrowseComponent implements OnInit, OnDestroy {
   @Input()
   numCols = 4;
 
-  constructor(public optionsService: OptionsService, private media: MediaObserver,
+  constructor(private media: MediaObserver,
     private layoutService: LayoutService) {
   }
 
@@ -32,6 +33,45 @@ export class BrowseComponent implements OnInit, OnDestroy {
     this.mediaSub = this.media.media$.subscribe((change: MediaChange) => {
       this.updateCols(change.mqAlias);
     });
+
+    enum OptionType {
+      ResearchActivity = 1,
+      Category,
+      Menu
+    }
+
+    enum CategoryId {
+      All = 1,
+      Support,
+      Equipment,
+      Training,
+      Software,
+      Facilities,
+      Guide,
+      Person,
+      Policies,
+      Articles,
+      SubHubs
+    }
+
+    this.categoryOptions = [
+      { id: CategoryId.All, name: 'All Categories', icon: 'public', type: OptionType.Category },
+      { id: CategoryId.Support, name: 'Service', icon: 'local_play', type: OptionType.Category },
+      { id: CategoryId.Equipment, name: 'Equipment', icon: 'build', type: OptionType.Category },
+      { id: CategoryId.Training, name: 'Training', icon: 'school', type: OptionType.Category },
+      { id: CategoryId.Software, name: 'Software', icon: 'desktop_mac', type: OptionType.Category },
+      { id: CategoryId.Facilities, name: 'Facility', icon: 'home', type: OptionType.Category },
+      { id: CategoryId.Guide, name: 'Guide', icon: 'import_contacts', type: OptionType.Category },
+      { id: CategoryId.Person, name: 'People', icon: 'face', type: OptionType.Category },
+      { id: CategoryId.Policies, name: 'Policy', icon: 'gavel', type: OptionType.Category },
+    ];
+
+    this.categoryOptionsGQL = [
+      { id: CategoryId.All, name: 'All Content', icon: 'public', type: OptionType.Category, url: '/all' },
+      { id: CategoryId.Equipment, name: 'Equipment', icon: 'build', type: OptionType.Category, url: '/equipment' },
+      { id: CategoryId.Articles, name: 'Articles', icon: 'import_contacts', type: OptionType.Category, url: '/articles' },
+      { id: CategoryId.SubHubs, name: 'SubHubs', icon: 'build', type: OptionType.Category, url: '/subhubs' }
+    ];
   }
 
   updateCols(mqAlias: string) {
