+++ research-hub-web/src/app/components/home/research-activity/research-activity.component.html
@@ -3,7 +3,7 @@
 <mat-chip-list>
   <div fxLayout="row wrap" fxLayout.lt-sm="column" fxLayoutAlign="space-between" fxLayoutAlign.lt-lg="flex-start"
     style="width: 100%">
-    <div *ngFor="let activity of optionsService.researchActivityOptions" [ngClass]="activity.className">
+    <div *ngFor="let activity of researchActivityOptions" [ngClass]="activity.className">
       <!-- Need to specify class for containing div also in order to use custom themes on material components -->
       <mat-chip selected [routerLink]="['/search']" [queryParams]="getQueryParams(activity)"
         [ngClass]="[activity.className, 'ra-chip']" [ngClass.lt-md]="[activity.className, 'ra-chip', 'tablet-view']"
