+++ research-hub-web/src/app/components/search-results/filter-dialog/filter-dialog.component.ts
@@ -1,8 +1,21 @@
 import { Component, Inject, Input } from '@angular/core';
 import { FormControl, FormGroup } from '@angular/forms';
-import { CategoryId } from '@services/options.service';
 import { SearchFiltersService, DEFAULT_FILTERS_VALUE } from '../search-filters/search-filters.service';
 import { AnalyticsService } from '@services/analytics.service';
+
+enum CategoryId {
+  All = 1,
+  Support,
+  Equipment,
+  Training,
+  Software,
+  Facilities,
+  Guide,
+  Person,
+  Policies,
+  Articles,
+  SubHubs
+}
 
 @Component({
   selector: 'app-filter-dialog',
