+++ research-hub-web/src/app/components/search-results/filter-sidenav/filter-sidenav.component.spec.ts
@@ -5,7 +5,6 @@ import { SearchResultsModule } from '../search-results.module';
 import { SearchFiltersService } from '../search-filters/search-filters.service';
 import { ResearchHubApiService } from '@services/research-hub-api.service';
 import { HttpClient, HttpHandler } from '@angular/common/http';
-import { OptionsService } from '@services/options.service';
 import { CommonModule } from '@angular/common';
 import { SharedModule } from '@components/shared/app.shared.module';
 import { SearchResultsComponentService } from '../search-results-component.service';
@@ -24,7 +23,6 @@ describe('FilterSidenavComponent', () => {
         SearchResultsComponentService,
         SearchFiltersService,
         ResearchHubApiService,
-        OptionsService,
         HttpClient,
         HttpHandler
       ]
