+++ research-hub-web/src/app/components/search-results/search-filters/search-filters.component.ts
@@ -2,12 +2,47 @@ import { Component, Input, OnDestroy, OnInit, ViewEncapsulation } from '@angular
 import { FormGroup } from '@angular/forms';
 import { Subscription, Observable, forkJoin } from 'rxjs';
 import { Tag } from '../mat-tags/mat-tags.component';
-import { OptionsService, RoleTypeId } from '@services/options.service';
 import { ResearchHubApiService, PeopleParams, Params } from '@services/research-hub-api.service';
 import { SearchResultsComponent } from '../search-results.component';
 import { ListItem } from '@model/ListItem';
 import { OrgUnit } from '@model/OrgUnit';
 import { AnalyticsService } from '@services/analytics.service';
+
+enum RoleTypeId {
+  ServiceOwner = 1,
+  CopyOwner,
+  UserSupport
+}
+
+enum OptionType {
+  ResearchActivity = 1,
+  Category,
+  Menu
+}
+
+enum CategoryId {
+  All = 1,
+  Support,
+  Equipment,
+  Training,
+  Software,
+  Facilities,
+  Guide,
+  Person,
+  Policies,
+  Articles,
+  SubHubs
+}
+
+enum ContentTypeId {
+  Support = 1,
+  Equipment,
+  Training,
+  Software,
+  Facilities,
+  KnowledgeArticle,
+  Guide
+}
 
 @Component({
   selector: 'app-search-filters',
@@ -16,7 +51,8 @@ import { AnalyticsService } from '@services/analytics.service';
   encapsulation: ViewEncapsulation.None
 })
 export class SearchFiltersComponent implements OnInit, OnDestroy {
-
+  public categoryOptions: any[];
+  public contentTypeMap: any;
   private searchCatSub: Subscription;
   private dataSub: Subscription;
 
@@ -32,7 +68,26 @@ export class SearchFiltersComponent implements OnInit, OnDestroy {
   public personTagSource: Tag[] = [];
   public orgUnitTagSource: Tag[] = [];
 
-  constructor(private apiService: ResearchHubApiService, public optionsService: OptionsService, public analyticsService: AnalyticsService) {
+  constructor(private apiService: ResearchHubApiService, public analyticsService: AnalyticsService) {
+    this.contentTypeMap = {};
+    this.contentTypeMap[CategoryId.Support] = [ContentTypeId.Support];
+    this.contentTypeMap[CategoryId.Equipment] = [ContentTypeId.Equipment];
+    this.contentTypeMap[CategoryId.Training] = [ContentTypeId.Training];
+    this.contentTypeMap[CategoryId.Software] = [ContentTypeId.Software];
+    this.contentTypeMap[CategoryId.Facilities] = [ContentTypeId.Facilities];
+    this.contentTypeMap[CategoryId.Guide] = [ContentTypeId.Guide, ContentTypeId.KnowledgeArticle];
+
+    this.categoryOptions = [
+      { id: CategoryId.All, name: 'All Categories', icon: 'public', type: OptionType.Category },
+      { id: CategoryId.Support, name: 'Service', icon: 'local_play', type: OptionType.Category },
+      { id: CategoryId.Equipment, name: 'Equipment', icon: 'build', type: OptionType.Category },
+      { id: CategoryId.Training, name: 'Training', icon: 'school', type: OptionType.Category },
+      { id: CategoryId.Software, name: 'Software', icon: 'desktop_mac', type: OptionType.Category },
+      { id: CategoryId.Facilities, name: 'Facility', icon: 'home', type: OptionType.Category },
+      { id: CategoryId.Guide, name: 'Guide', icon: 'import_contacts', type: OptionType.Category },
+      { id: CategoryId.Person, name: 'People', icon: 'face', type: OptionType.Category },
+      { id: CategoryId.Policies, name: 'Policy', icon: 'gavel', type: OptionType.Category },
+    ];
   }
 
   ngOnInit() {
