+++ research-hub-web/src/app/components/search-results/search-results.component.html
@@ -1,7 +1,5 @@
 <div #resultsDummyHeader></div>
-
 <h2 class="search-results-title">Results</h2>
-
 <div class="summary-section" *ngIf="!showEmptyState">
   <div fxLayout="row" fxLayoutAlign="start center">
     <p fxFlex [innerHTML]="resultsSummary" style="margin-bottom: 0; margin-right: 1em;"></p>
