+++ research-hub-web/src/app/services/services.module.ts
@@ -3,7 +3,6 @@ import { CommonModule } from '@angular/common';
 import { HttpClientModule } from '@angular/common/http';
 import { AnalyticsService } from './analytics.service';
 import { ResearchHubApiService } from './research-hub-api.service';
-import { OptionsService } from './options.service';
 import { LayoutService } from './layout.service';
 import { CerApiService } from './cer-api.service';
 
@@ -18,7 +17,6 @@ import { CerApiService } from './cer-api.service';
     AnalyticsService,
     CerApiService,
     ResearchHubApiService,
-    OptionsService,
     LayoutService
   ],
   exports: []
